Fix type inference bug in the new quant signal history query
Production hit this immediately: WHERE as_of_date >= (subquery) - \$2
let Postgres infer \$2 as date (from the subtraction), making the
subtraction's result an integer (date - date) instead of a date, so
the outer >= comparison failed with "operator does not exist: date >=
integer". Explicit \$2::integer cast fixes the inference.

+++ web/backend/routers/signals.py
@@ -219,7 +219,7 @@ async def quant_signal_history(
             """
             SELECT as_of_date, signal, expected_return_pct, target_price, last_close
             FROM pit_quant_signal
-            WHERE ticker = $1 AND as_of_date >= (SELECT max(as_of_date) FROM pit_quant_signal) - $2
+            WHERE ticker = $1 AND as_of_date >= (SELECT max(as_of_date) FROM pit_quant_signal) - $2::integer
             ORDER BY as_of_date ASC
             """,
             ticker.strip().upper(), days,
